# swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Hyperledger Fabric Gateway API"
host: "localhost:4000"
basePath: "/"
schemes:
- "http"
produces:
- "application/json"
paths:
...
  /chaincode/install:
    post:
      tags:
      - "Chaincode"
      summary: "Install chaincode"
      description: "Install chaincode on the channel at all endorse peers"
      operationId: "chaincodeInstall"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "install request"
        required: true
        schema:
          type: "object"
          required: ["chaincodeName","channelName","chaincodePath","chaincodeVersion","sourceType","user"]
          properties:
            chaincodeName:
              example: mycc
              type: "string"
            channelName:
              type: "string"
              example: mychannel
              # required: true
            chaincodePath:
              type: "string"
              example: chaincodes/testCC
              # required: true
            chaincodeVersion:
              type: "string"
              example: v1
              # required: true
            sourceType:
              enum: ["sourceCode", "package"]
              type: "string"
              example: sourceCode
            user:
              $ref: "#/definitions/user"
              # required: true
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
...
  /chaincode/instantiate:
    post:
      tags:
      - "Chaincode"
      summary: "instantiate chaincdoe"
      description: "instantiate chaincode at all endorse peers of a channel, chaincode should be install chaincode on peer first"
      operationId: "chaincodeInstantiate"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "Instantiate request"
        required: true
        schema:
          type: "object"
          required: [
        "channelName",
        "chaincodeName",
        "chaincodeVersion",
        "functionName",
        "args",
        "user"
    ]
          properties:
            chaincodeName:
              type: "string"
              example: mycc
            channelName:
              type: "string"
              example: mychannel
            chaincodeVersion:
              type: "string"
              example: v1
            functionName:
              type: "string"
              example: init
            args:
              type: "array"
              example: ["a","100","b","200"]
              items:
                type: "string"
            opt:
              type: "object"
              properties:
                endorsement-policy:
                  type: string
                  $ref: "#/definitions/endorsement-policy"
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/upgrade:
    post:
      tags:
      - "Chaincode"
      summary: "upgrade chaincdoe"
      description: "upgrade chaincode at all endorse peer of a channel, chaincode should be install chaincode on peer first"
      operationId: "chaincodeUpgrade"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "upgrade request"
        required: true
        schema:
          type: "object"
          required: [
        "channelName",
        "chaincodeName",
        "chaincodeVersion",
        "functionName",
        "args",
        "user"
    ]
          properties:
            chaincodeName:
              type: "string"
              example: mycc
            channelName:
              type: "string"
              example: mychannel
            chaincodeVersion:
              type: "string"
              example: v1
            functionName:
              type: "string"
              example: init
            args:
              type: "string"
              example: ["a","1000","b","200"]
            user:
              $ref: "#/definitions/user"
            opt:
              type: "object"
              properties:
                endorsement-policy:
                  type: string
                  $ref: "#/definitions/endorsement-policy"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/invoke:
    post:
      tags:
      - "Chaincode"
      summary: "invoke chaincdoe"
      description: "Send endorse request to all endorse peer on the channel at one time, then send the reponse to order if there any response is good, it may cause rejection of insufficiency endorsement"
      operationId: "chaincodeInvoke"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "Instantiate request"
        required: true
        schema:
          type: "object"
          required: ["channelName", "chaincodeName", "functionName", "args", "user"]
          properties:
            chaincodeName:
              type: "string"
              example: mycc
            channelName:
              type: "string"
              example: mychannel
            functionName:
              type: "string"
              example: invoke
            args:
              type: "array"
              example: ["move","a","b","10"]
              items:
                type: "string"
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/invokeE:
    post:
      tags:
      - "Chaincode"
      summary: "invoke chaincdoe with endorsepolicy"
      description: "Different with \"\/chaincode\/invoke\", this rest api will handle the endorse request based on endorse policy, to invoke this api, you should make sure your monitr module has been enabled"
      operationId: "chaincodeInvokeE"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "Instantiate request"
        required: true
        schema:
          type: "object"
          required: ["channelName", "chaincodeName", "functionName", "args", "user"]
          properties:
            chaincodeName:
              type: "string"
              example: mycc
            channelName:
              type: "string"
              example: mychannel
            functionName:
              type: "string"
              example: invoke
            args:
              type: "array"
              example: ["move","a","b","10"]
              items:
                type: "string"
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/query:
    post:
      tags:
      - "Chaincode"
      summary: "query chaincode"
      description: "Send qeury request to all peer of self organiztion, then compare each result, if any result is different from others, the gateway will send back a array contained all peer's responses, if not, gateway will just send a resonpse back"
      operationId: "chaincodeQuery"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "Instantiate request"
        required: true
        schema:
          type: "object"
          required: ["channelName", "chaincodeName", "functionName", "args", "user"]
          properties:
            chaincodeName:
              type: "string"
              example: mycc
            channelName:
              type: "string"
              example: mychannel
            functionName:
              type: "string"
              example: invoke
            args:
              type: "array"
              example: ["query","a"]
              items:
                type: "string"
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/queryHistory:
    post:
      tags:
      - "Chaincode"
      summary: "query chaincdoe"
      description: "Send qeury request to all peer of self organiztion, then compare each result, if any result is different from others, the gateway will send back a array contained all peer's responses, if not, gateway will just send a resonpse back"
      operationId: "chaincodeQueryHistory"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "query history request"
        required: true
        schema:
          type: "object"
          required: ["channelName", "chaincodeName", "functionName", "args", "user"]
          properties:
            chaincodeName:
              type: "string"
              example: "mycc"
            channelName:
              type: "string"
              example: "mychannel"
            functionName:
              type: "string"
              example: "query"
            args:
              type: "array"
              example: ["a"]
              items:
                type: "string"
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
  /chaincode/installed:
    post:
      tags:
      - "Chaincode"
      summary: "query installed chaincdoe"
      description: "get installed chaincode from a self organiztion's peer, peer's name and url was mapped on the network-config.json"
      operationId: "getInstalledChaincodes"
      produces:
      - "application/json"
      parameters:
      - name: "request"
        in: "body"
        description: "query history request"
        required: true
        schema:
          type: "object"
          required: ["peerName", "user"]
          properties:
            peerName:
              type: "string"
              example: peer1
            user:
              $ref: "#/definitions/user"
      responses:
        200:
          description: "SDK response"
          schema:
            $ref: "#/definitions/sdkResoponse"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/inline_response_default"
      x-swagger-router-controller: "chaincode"
...
definitions:
...
  inline_response_default:
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"
...
  sdkResoponse:
    properties:
      sdkResult:
        type: "object"
  user:
    required: [enrollID,enrollSecret]
    properties:
        enrollID:
          type: "string"
          example: orgAdmin
        enrollSecret:
          type: "string"
          example: "87654321"
...
  endorsement-policy:
    properties:
      identities:
        type: array
        items:
          type: "object"
          properties:
            role:
              type: object
              properties:
                name:
                  type: string
                  enum: [member,admin]
                  example: member
                mspId:
                  type: string
                  example: Org2MSP
      policy:
        type: object
        example:
          1-of: [
            signed-by: 0
            ]
securityDefinitions:
   # X-API-Key: abcdef12345
   APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key